+++ src/compliment/core.clj
@@ -6,7 +6,7 @@
 (ns compliment.core
   "Core namespace. Most interactions with Compliment should happen
   through functions defined here."
-  (:require (compliment.sources ns-mappings
+  (:require (compliment.sources vars
                                 namespaces
                                 classes
                                 class-members